fix(472): Add padding to the dialogue box (#475)
* Add padding to the dialogue box

* Fix TextCanAppearInstantly test

I'm assuming that since the dialogue box has changed size and will only allow 3 lines instead of 4, that caused this long line of test text to fail the test. I changed the test line to what the test expected. If there's a better way to handle this please teach me the ways.